% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedDataPrep.R
\name{Data}
\alias{Data}
\title{A simple data generation method for use benchmarking}
\usage{
Data(nx, beta = c(0.4, 0.6, -0.4), phi = 1, probset1 = c(0.4, 0.3, 0.3),
  probset2 = c(0.1, 0.2, 0.5))
}
\arguments{
\item{nx}{Number of variants.}

\item{phi}{Covariance of random effects (genetic additive variance).}

\item{probset1}{Probabilities of genotypes for first half of dataset ie (c(0.1,0.2,0.5)) .}

\item{probset2}{Probabilities of genotypes for second half of dataset ie (c(0.1,0.2,0.5)) .}

\item{X.re}{Design matrix of random effects.}

\item{c}{Covariance of fixed effects.}
}
\value{
A list of phenotypes ($y), random effects ($X.re), fixed effects ($X.fe), probabilites
for the phenotype ($probs).
}
\description{
Generate a dataset with 3 fixed effects and with an arbitrary number
of genotypes and some relatedness.
}
