% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitBenchmarkData.R
\name{generateBenchmarks}
\alias{generateBenchmarks}
\title{A simple data generation method for use benchmarking}
\usage{
generateBenchmarks(data = system.file("extdata", "sampleData.tsv", package =
  "hailLogitMMSupport"), neff = 3, phi = 0.007, c = 1)
}
\arguments{
\item{data}{a datset for benchmarking.}

\item{neff}{number of fixed effects}

\item{phi}{Covariance of random effects (genetic additive variance).}

\item{c}{Covariance of fixed effects.}
}
\value{
A data frame of fixed effect parameters benched with four aforementioned models.
}
\description{
A general file for simple test and use cases for 
all implemented versions of logistic (mixed) models1) GLM - standard classic logistic regression R implementation for fixed effect estimate benchmarks
2) logit.EM.R - EM version of logistic regression with Polya-Gamma latent variables
3) logit.PG.EM.mm.gen.gibbs - Gibbs sampling version of logistic mixed model regression with Polya-Gamma latents
4) logit.PG.EM.mm.gen - EM version of logistic mixed model regression with Polya-Gamma latents

The sample data has been created from Balding-Nichols model
and was exported as a file as given in Jupyter Notebook sampleNotebook.ipynb.
3 fixed effects (V1,V2,V3) were extracted from sampleAnnot.tsv.
Phenotype column Pheno.
}
\examples{
generateBenchmarks(system.file('extdata','sampleData.tsv',package='hailLogitMMSupport'))

Dat=Data(200)
generateBenchmarks(Dat)

}
