% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitFunctions.R
\name{logit.PG.EM.fastlmm.ksisamp}
\alias{logit.PG.EM.fastlmm.ksisamp}
\title{Alternative EM-based logistic mixed effect regression on Polya-Gamma distributed latent variables.}
\usage{
logit.PG.EM.fastlmm.ksisamp(y, X.re, X.fe, phi, c = 1, n = rep(1,
  length(y)), m0 = rep(0, ncol(X.fe)), P0 = matrix(0, nrow = ncol(X.fe),
  ncol = ncol(X.fe)))
}
\arguments{
\item{y}{A boolean phenotype.}

\item{X.re}{Design matrix of random effects.}

\item{X.fe}{Design matrix of fixed effects.}

\item{phi}{Covariance of random effects (genetic additive variance).}

\item{c}{Covariance of fixed effects.}
}
\value{
Logistic regression parameter posteriors for the fixed effect coefficients
specified in \code{X.fe}.
}
\description{
Performs logistic mixed effect regression on fixed effects \code{X.fe} and \code{X.re}.
Algorithm employs Polya-Gamma latent variables to transform the phenotype into Gaussian framework.
Parameter coefficients are from iterative adjustments to Polya-Gamma omega (latent variables),
beta (fixed effects) and gamma (random effects).
Omega expected values come from expecepted values of Polya-Gamma distribution, beta and omega
values use maximisation of Gaussian parameters from the maximum likelihood estimate
of Bayesian linear regression. Note! It is extremely slow and not included in benchmarking.
}
