% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitFunctions.R
\name{logit.PG.EM.mm.gen.gibbs}
\alias{logit.PG.EM.mm.gen.gibbs}
\title{Logistic mixed effect regression with Gibbs sampling algorithm based on Polya-Gamma distributed latent variables.}
\usage{
logit.PG.EM.mm.gen.gibbs(y, X.re, X.fe, phi, c, inc.m = F, n = rep(1,
  length(y)), m0 = rep(0, ncol(X.fe)), kappa = 0, samp = 500,
  burn = 500, verbose = 100, P0 = matrix(0, nrow = ncol(X.fe), ncol =
  ncol(X.fe)))
}
\arguments{
\item{y}{A boolean phenotype.}

\item{X.re}{Design matrix of random effects.}

\item{X.fe}{Design matrix of fixed effects.}

\item{phi}{Covariance of random effects (genetic additive variance).}

\item{c}{Covariance of fixed effects.}

\item{inc.m}{boolean for whether the intercept should be estimated?}
}
\value{
Logistic regression parameter posteriors for the fixed effect coefficients
specified in \code{X.fe}.
}
\description{
Performs logistic mixed effect regression on fixed effects \code{X.fe} and \code{X.re}.
Algorithm employs Polya-Gamma latent variables to transform the phenotype into Gaussian framework.
Parameter coefficients are estimated using a Gibbs sampler that returns a posterior for 
estimated fixed effects. Adapted from mixed-effect model in BayesLogit package. Point estimates for
the mixed effect estimates could be retrieved from posterior means.
}
