% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitFunctions.R
\name{logit.PG.EM.mm.gen}
\alias{logit.PG.EM.mm.gen}
\title{Logistic mixed effect regression with EM algorithm based on Polya-Gamma distributed latent variables.}
\usage{
logit.PG.EM.mm.gen(y, X.re, X.fe, phi, c, inc.m = F, n = rep(1, length(y)),
  m0 = rep(0, ncol(X.fe)))
}
\arguments{
\item{y}{A boolean phenotype.}

\item{X.re}{Design matrix of random effects.}

\item{X.fe}{Design matrix of fixed effects.}

\item{phi}{Covariance of random effects (genetic additive variance).
@param inc.m boolean for whether the intercept should be estimated?}

\item{c}{Covariance of fixed effects.}
}
\value{
Logistic regression parameters for the fixed effects and random effects
specified in \code{X.re} and \code{X.fe}.
}
\description{
Performs logistic mixed effect regression on fixed effects \code{X.fe} and \code{X.re}.
Algorithm employs Polya-Gamma latent variables to transform the phenotype into Gaussian framework.
Parameter coefficients are estimated using an expectation-maximisation algorithm.
}
